cli: replace cylc scan with new implementaion
