+++ cylc/flow/scripts/cylc_scan.py
@@ -15,39 +15,104 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""cylc [discovery] scan [OPTIONS]
 
-"""cylc [discovery] scan [OPTIONS] [HOSTS ...]
+List Cylc workflows, by default this shows only running or held workflows.
 
-Print information about running suites.
+Examples:
+    # list all "active" workflows (i.e. running or held)
+    $ cylc scan
 
-Use the -o/--suite-owner option to get information of running suites for other
-users.
+    # show more information about these workfows
+    $ cylc scan -t rich
 
-Titles, descriptions, state totals, and cycle point state totals may also be
-revealed publicly, depending on global and suite authentication settings.
+    # don't rely on colour for job state totals
+    $ cylc scan -t rich --colour-blind
 
-WARNING: a suite suspended with Ctrl-Z will cause port scans to hang until the
-connection times out (see --comms-timeout)."""
+    # list all "inactive" workflows (i.e. registered or stopped)
+    $ cylc scan --state stopped
 
+    # list all workflows (active or inactive)
+    $ cylc scan --state=running,held,stopped
+    $ cylc scan --state=all  # or using the shorthand
+
+    # filter workflows by name
+    $ cylc scan --name '^f.*'  # show only flows starting with "f"
+
+    # get results in JSON format
+    $ cylc scan -t json
+"""
+
+import asyncio
 import json
+from pathlib import Path
 
-from colorama import Fore, Style, init as color_init
+from ansimarkup import ansiprint as cprint
 
-from cylc.flow.exceptions import UserInputError
+from cylc.flow import LOG
 from cylc.flow.network.scan import (
-    get_scan_items_from_fs, re_compile_filters, scan_many)
+    scan,
+    is_active,
+    contact_info,
+    graphql_query,
+    filter_name
+)
 from cylc.flow.option_parsers import CylcOptionParser as COP
-from cylc.flow.suite_status import KEY_META, KEY_NAME, KEY_OWNER, KEY_STATES
-from cylc.flow.task_state import TASK_STATUSES_ORDERED
-from cylc.flow.task_state_prop import get_status_prop
-from cylc.flow import terminal
+from cylc.flow.print_tree import print_tree
+from cylc.flow.suite_files import ContactFileFields as Cont
+from cylc.flow.terminal import cli_function
 
 
-INDENT = "   "
-TITLE_STYLE = Style.BRIGHT
-MISSING_STYLE = Fore.MAGENTA
-ERROR_STYLE = Fore.RED
-META_KEY_ORDER = ['title', 'description', 'group']
+# all supported suite states
+FLOW_STATES = {
+    'running',
+    'held',
+    'stopped'
+}
+
+
+# suite status colours
+FLOW_STATE_CMAP = {
+    # suite state: term colour
+    'running': 'green',
+    'held': 'fg 172',
+    'stopped': 'red'
+}
+
+
+# suite status symbols
+FLOW_STATE_SYMBOLS = {
+    # NOTE: the standard media control characters ▶️,, ⏸️,, ⏹️
+    #       can appear wildly different font-depending and may not
+    #       even be monospace
+    'running': '▶',
+    'held': '‖',
+    'stopped': '■'
+}
+
+
+# graphql fields to request from the workflow for the "rich" format
+RICH_FIELDS = {
+    'status': None,
+    'stateTotals': None,
+    'cylcVersion': None,
+    'meta': {
+        'title',
+        'description',
+    },
+
+}
+
+
+# job icon colours
+JOB_COLOURS = {
+    # job status: term colour
+    'submitted': 'fg 38',
+    'submit-failed': 'fg 13',
+    'running': 'fg 27',
+    'succeeded': 'fg 34',
+    'failed': 'fg 124'
+}
 
 
 def get_option_parser():
@@ -59,267 +124,309 @@ def get_option_parser():
     )
 
     parser.add_option(
-        '--ordered',
-        help='Display results in order, this may take longer.',
-        action='store_true')
-
-    parser.add_option(
-        "-n", "--name",
-        metavar="PATTERN",
-        help="List suites with name matching PATTERN (regular expression). "
-             "Defaults to any name. Can be used multiple times.",
-        action="append", dest="patterns_name", default=[])
-
-    parser.add_option(
-        "-o", "--suite-owner",
-        metavar="PATTERN",
-        help="List suites with owner matching PATTERN (regular expression). "
-             "Defaults to current user. Use '.*' to match all known users. "
-             "Can be used multiple times.",
-        action="append", dest="patterns_owner", default=[])
-
-    parser.add_option(
-        "-d", "--describe",
-        help="Print suite metadata if available.",
-        action="store_true", default=False, dest="describe")
-
-    parser.add_option(
-        "-s", "--state-totals",
-        help="Print number of tasks in each state if available "
-             "(total, and by cycle point).",
-        action="store_true", default=False, dest="state_totals")
-
-    parser.add_option(
-        "--publisher",
-        help="Append the suite publisher information to output.",
-        action="store_true", default=False, dest="publisher")
-
-    parser.add_option(
-        "-f", "--full",
-        help="Print all available information about each suite.",
-        action="store_true", default=False, dest="full")
-
-    parser.add_option(
-        '--color', '--colour', action='store', dest='color', default='auto',
-        help='Colorize the output, can be "always", "never" or "auto".')
-
-    parser.add_option(
-        "--comms-timeout", metavar="SEC",
-        help="Set a timeout for network connections "
-             "to each running suite. The default is 5 seconds.",
-        action="store", default=5.0, dest="comms_timeout")
-
-    parser.add_option(
-        '-t',
-        '--format',
-        default='plain',
-        action='store',
-        dest='format',
+        '--name',
         help=(
-             'Set output format:\n'
-             ' * plain (default) - text format for interactive use\n'
-             ' * raw - parsable format (suite|owner|host|property|value)\n'
-             ' * json - JSON format ({suite: {owner: OWNER, host: HOST ...)')
+            'Filter flows by registered name using a regex.'
+            ' Can be used multiple times, workflows will be displayed if'
+            ' their name matches ANY of the provided regexes.'
+        ),
+        action='append'
+    )
+
+    parser.add_option(
+        '--states',
+        help='Choose which flows to display by providing a list of states.',
+        default='running,held',
+        action='store'
+    )
+
+    parser.add_option(
+        '--format', '-t',
+        help='Set the output format:',
+        choices=('rich', 'plain', 'json', 'tree'),
+        default='plain'
+    )
+
+    parser.add_option(
+        '--sort',
+        help='Sort flows by name before writing them (takes longer).',
+        action='store_true'
+    )
+
+    parser.add_option(
+        '--colour-blind', '--color-blind',
+        help=(
+            "Don't depend on colour to convey information. "
+            ' Use this rather than --color=never so you still get bold text.'
+        ),
+        action='store_true'
     )
 
     return parser
 
 
-@terminal.cli_function(get_option_parser)
-def main(parser, options):
-    """Implement "cylc scan"."""
-    if options.full:
-        options.describe = options.state_totals = options.publisher = True
-    if options.format in ['raw', 'json']:
-        options.color = False
+def state_totals(totals, colour_blind=False):
+    """Return a string with a visual representation of a suite's state totals.
 
-    # color settings
-    if options.color in ['auto', 'always'] and terminal.supports_color():
-        options.color = True
-    else:
-        options.color = False
-    color_init(autoreset=True, strip=not options.color)
+    Args:
+        totals (dict):
+            State totals dictionary.
+        colour_blind (bool):
+            If True then we will not depend on colour to convey information.
 
-    # name and owner patterns
-    if options.patterns_name:
-        patterns_name = options.patterns_name
-    else:
-        patterns_name = ['.*']  # Any suite name.
-    patterns_owner = None
-    if options.patterns_owner:
-        patterns_owner = options.patterns_owner
-    try:  # Compile and check "name" and "owner" regular expressions
-        cre_owner, cre_name = re_compile_filters(patterns_owner, patterns_name)
-    except ValueError as exc:
-        parser.error(str(exc))
+    Returns:
+        str
 
-    # list of endpoints to call
-    methods = ['identify']
-    if options.describe:
-        methods.append('describe')
-    if options.state_totals:
-        methods.append('state_totals')
+    """
+    ret = []
+    for state, tag in JOB_COLOURS.items():
+        number = totals[state]
+        if number == 0:
+            continue
+        if colour_blind:
+            ret.append(f'<{tag}>{state}:{number}</{tag}>')
+        else:
+            ret.append(f'<{tag}>{number} \u25A0</{tag}>')
+    return ' '.join(ret)
 
-    # suite generator
-    suites = scan_many(
-        get_scan_items_from_fs(cre_owner, cre_name),
-        timeout=options.comms_timeout,
-        methods=methods,
-        ordered=options.ordered
+
+def state_totals_key():
+    """Return a key to accompany state_totals."""
+    return '<b>Job State Key: </b>' + ' '.join(
+        f'<{tag}>\u25A0 {name}</{tag}>'
+        for name, tag in JOB_COLOURS.items()
     )
 
-    # determine output format
-    if options.format == 'json':
-        print(json.dumps(list(suites), indent=4))
-    elif options.format == 'raw':
-        formatter = format_raw
-    elif options.format == 'plain':
-        formatter = format_plain
+
+def _format_json(items, _):
+    """A JSON formatter."""
+    return json.dumps(
+        [
+            {
+                key: str(value) if isinstance(value, Path) else value
+                for key, value in flow.items()
+            }
+            for flow in items
+        ],
+        indent=4,
+        sort_keys=True
+    )
+
+
+def _format_plain(flow, _):
+    """A single line format of the form: <name> [<host>:<port>]"""
+    if flow.get('contact'):
+        return f'<b>{flow["name"]}</b> {flow[Cont.HOST]}:{flow[Cont.PORT]}'
     else:
-        raise UserInputError('Unknown format: %s' % options.format)
-
-    # output state legend if necessary
-    state_legend = ""
-    if options.color and options.state_totals:
-        n_states = len(TASK_STATUSES_ORDERED)
-        for index, state in enumerate(TASK_STATUSES_ORDERED):
-            state_legend += get_status_prop(state, 'ascii_ctrl')
-            if index == n_states / 2:
-                state_legend += "\n"
-        print(state_legend.rstrip() + "\n")
-
-    # work through scan results one by one
-    for reg, host, port, pub_port, api, info in suites:
-        if isinstance(info, str):
-            print(ERROR_STYLE + ' '.join([reg, host, port, info]))
-        elif info is None:
-            print(ERROR_STYLE +
-                  ' '.join([reg, host, port, 'Error Connecting']))
-        elif info[KEY_NAME] != reg:
-            # TODO - should we do anything here, is this likely?
-            print(ERROR_STYLE + 'Warning: suite has changed name %s => %s' % (
-                reg, info[KEY_NAME]))
-        else:
-            formatter(reg, host, port, pub_port, api, info, options)
+        return f'<dim><b>{flow["name"]}</b></dim>'
 
 
-def sort_meta(item):
-    """Sort meta items for the "plain" output format."""
-    key = item[0]
-    if key in META_KEY_ORDER:
-        return str(META_KEY_ORDER.index(key))
-    return key
-
-
-def format_plain(name, host, port, pub_port, api, info, options):
-    """Print a scan result, implements --format=plain"""
-    owner = info[KEY_OWNER]
-
-    if options.publisher:
-        print(Style.BRIGHT + name + Style.NORMAL
-              + ' %s@%s:%s' % (owner, host, port)
-              + ' %s@%s:%s' % (owner, host, pub_port))
+def _format_rich(flow, opts):
+    """A multiline format which pulls in metadata."""
+    status = flow.get('status', 'stopped')
+    if opts.colour_blind:
+        name = f'{flow["name"]} ({status})'
     else:
-        print(Style.BRIGHT + name + Style.NORMAL
-              + ' %s@%s:%s' % (owner, host, port))
+        symbol = FLOW_STATE_SYMBOLS[status]
+        tag = FLOW_STATE_CMAP[status]
+        name = f'<{tag}>{symbol}</{tag}> {flow["name"]}'
+    if not flow.get('contact') or 'status' not in flow:
+        ret = [f'<dim><b>{name}</b></dim>']
+    else:
+        ret = [f'<b>{name}</b>']
 
-    if options.describe:
-        meta_items = info.get(KEY_META)
-        meta_items['API'] = api
-        if meta_items is None:
-            print(INDENT + MISSING_STYLE + "(description withheld)")
-            return
-        for metaitem, metavalue in sorted(meta_items.items(), key=sort_meta):
-            if metaitem in META_KEY_ORDER:
-                metaitem = metaitem.capitalize()
-            print(INDENT + TITLE_STYLE + metaitem + ":")
-            if not metavalue:
-                metavalue = MISSING_STYLE + '(no %s)' % metaitem + Fore.RESET
-            for line in metavalue.splitlines():
-                print(INDENT * 2 + line)
-
-    if options.state_totals:
-        totals = info.get(KEY_STATES)
-        if totals is None:
-            print(INDENT + MISSING_STYLE + "(state totals withheld)")
-            return
-        print(INDENT + TITLE_STYLE + "Task state totals:")
-        for point, state_line in get_point_state_count_lines(
-                *totals, use_color=options.color):
-            point_prefix = ""
-            if point:
-                point_prefix = "%s " % point
-            print(INDENT * 2 + "%s%s" % (point_prefix, state_line))
+        display = {
+            'state totals': state_totals(
+                flow['stateTotals'], opts.colour_blind
+            ),
+            **{
+                key: flow['meta'][key] or '<dim>*null*</dim>'
+                for key in (
+                    'title',
+                    'description'
+                )
+            },
+            **{
+                name: flow[key]
+                for name, key in (
+                    ('cylcVersion', 'cylcVersion'),
+                    ('host', Cont.HOST),
+                    ('port', Cont.PORT)
+                )
+            }
+        }
+        maxlen = max(len(key) for key in display)
+        for key, value in display.items():
+            ret.append(f'    {key: <{maxlen}}   {value}')
+    return '\n'.join(ret)
 
 
-def format_raw(name, host, port, pub_port, api, info, options):
-    """Print a scan result, implements --format=raw"""
-    owner = info[KEY_OWNER]
+def sort_function(flow):
+    if flow.get('status') == 'running':
+        state = 0
+    elif flow.get('status') == 'held':
+        state = 0
+    elif flow.get('contact'):
+        state = 2
+    else:
+        state = 3
+    return (state, flow['name'])
 
-    if options.publisher:
-        print("%s|%s|%s|port|%s|publish-port|%s" % (
-            name, owner, host, port, pub_port)
+
+async def _sorted(pipe, formatter, opts):
+    """List all flows, sort, then print them individually."""
+    ret = []
+    async for item in pipe:
+        ret.append(item)
+    for flow in sorted(ret, key=sort_function):
+        cprint(formatter(flow, opts))
+
+
+async def _serial(pipe, formatter, opts):
+    """List all flows, then print them as one."""
+    ret = []
+    async for item in pipe:
+        ret.append(item)
+    cprint(formatter(ret, opts))
+
+
+async def _async(pipe, formatter, opts):
+    """List and print flows individually."""
+    async for flow in pipe:
+        cprint(formatter(flow, opts))
+
+
+async def _tree(pipe, formatter, opts):
+    """List all flows, sort, then print them as a tree."""
+    # get list of flows
+    ret = []
+    async for flow in pipe:
+        ret.append(flow)
+
+    # construct tree
+    tree = {}
+    for flow in sorted(ret, key=lambda f: f['name']):
+        parts = Path(flow['name']).parts
+        pointer = tree
+        for part in parts[:-1]:
+            if part not in pointer:
+                pointer[part] = {}
+            pointer = pointer[part]
+        flow['name'] = parts[-1]
+        item = formatter(flow, opts)
+        if len(parts) > 1:
+            item = f' {item}'
+        pointer[item] = ''
+
+    # print tree
+    ret = print_tree(tree, '', sort=False, use_unicode=True)
+    cprint('\n'.join(ret))
+
+
+def get_pipe(opts, formatter):
+    """Construct a pipe for listing flows."""
+    pipe = scan
+
+    show_running = 'running' in opts.states
+    show_held = 'held' in opts.states
+    show_active = show_running or show_held
+    # show_active = bool({'running', 'held'} & opts.states)
+    show_inactive = bool({'stopped'} & opts.states)
+
+    # filter by flow name
+    if opts.name:
+        pipe |= filter_name(*opts.name)
+
+    # filter by flow state  -  TODO: reduce this
+    if show_active and show_inactive:
+        pipe |= is_active(True, filter_stop=False)
+    elif show_active:
+        pipe |= is_active(True)
+    elif show_inactive:
+        pipe |= is_active(False)
+
+    # get contact file information
+    if show_active:
+        pipe |= contact_info
+
+    graphql_fields = {}
+    graphql_filters = set()
+
+    # filter held/running flows
+    if show_active and not (show_running and show_held):
+        graphql_fields['status'] = None
+        graphql_filters.add((('status',), tuple(opts.states)))
+
+    # get fancy data if requested
+    if formatter == _format_rich:
+        # graphql_fields['status'] = None
+        graphql_fields.update(RICH_FIELDS)
+
+    # add graphql queries / filters to the pipe
+    if graphql_fields:
+        pipe |= graphql_query(graphql_fields, filters=graphql_filters)
+
+    return pipe
+
+
+def get_formatter(opts):
+    """Return the appropriate formatter and method for the provided opts."""
+    formatter = None
+    method = None
+
+    if opts.format == 'plain':
+        formatter = _format_plain
+        method = _async
+    elif opts.format == 'json':
+        formatter = _format_json
+        method = _serial
+    elif opts.format == 'rich':
+        formatter = _format_rich
+        method = _async
+    elif opts.format == 'tree':
+        formatter = _format_plain
+        method = _tree
+    else:
+        raise NotImplementedError
+
+    if opts.sort and method == _serial:
+        raise ValueError(f'--sort is incompatible with --format {opts.format}')
+    elif opts.sort:
+        method = _sorted
+
+    return formatter, method
+
+
+async def scanner(opts):
+    """Print workflows to stdout."""
+    formatter, method = get_formatter(opts)
+    pipe = get_pipe(opts, formatter)
+
+    LOG.debug(f'pipe: {repr(pipe)}')
+
+    await method(pipe, formatter, opts)
+
+
+@cli_function(get_option_parser)
+def main(parser, opts, color):
+    """Implement `cylc scan`."""
+    # validate / standardise the list of workflow states
+    opts.states = set(opts.states.split(','))
+    if 'all' in opts.states:
+        opts.states = FLOW_STATES
+    else:
+        assert opts.states
+        assert all(
+            state in FLOW_STATES
+            for state in opts.states
         )
-    else:
-        print("%s|%s|%s|port|%s" % (name, owner, host, port))
 
-    if options.describe:
-        # Extracting required data for these options before processing
-        meta_items = info.get(KEY_META)
-        meta_items['API'] = api
+    if not color:
+        # we cannot support colour or have been requested not to use it
+        opts.colour_blind = True
 
-        # clean_meta_items = {}
-        # for key, value in meta_items.items():
-        #     if value:
-        #         clean_meta_items.update({
-        #             key: ' '.join([x.strip() for x in
-        #                            str(value).split('\n') if x])})
+    # print state totals key as needed
+    if opts.format == 'rich' and not opts.colour_blind:
+        cprint(state_totals_key() + '\n')
 
-        # for key, value in meta_items.items():
-        #     if value:
-        #         clean_meta_items.update({
-        #             key: ' '.join([x.strip() for x in
-        #                            str(value).split('\n') if x])})
-
-        for key, value in sorted(meta_items.items(), key=sort_meta):
-            value = ' '.join(line.strip()
-                             for line in str(value).splitlines()
-                             if line)
-            print("%s|%s|%s|%s|%s" % (name, owner, host, key, value))
-
-    if options.state_totals:
-        totals = info.get(KEY_STATES)
-        if totals is None:
-            return
-        for point, state_line in get_point_state_count_lines(*totals):
-            key = KEY_STATES
-            if point:
-                key = "%s:%s" % (KEY_STATES, point)
-            print("%s|%s|%s|%s|%s" % (name, owner, host, key, state_line))
-
-
-def get_point_state_count_lines(state_count_totals, state_count_cycles,
-                                use_color=False):
-    """Yield (point, state_summary_text) tuples."""
-    line = ""
-    for state, tot in sorted(state_count_totals.items()):
-        if use_color:
-            subst = " %d " % tot
-            line += get_status_prop(state, 'ascii_ctrl', subst)
-        else:
-            line += '%s:%d ' % (state, tot)
-    yield ("", line.strip())
-
-    for point_string, state_count_cycle in sorted(state_count_cycles.items()):
-        line = ""
-        for state, tot in sorted(state_count_cycle.items()):
-            if use_color:
-                subst = " %d " % tot
-                line += get_status_prop(state, 'ascii_ctrl', subst)
-            else:
-                line += '%s:%d ' % (state, tot)
-        yield (point_string, line.strip())
-
-
-if __name__ == "__main__":
-    main()
+    asyncio.run(scanner(opts))
