+++ setup.cfg
@@ -105,7 +105,6 @@ console_scripts =
     cylc-restart = cylc.flow.scripts.cylc_restart:main
     cylc-run = cylc.flow.scripts.cylc_run:main
     cylc-scan = cylc.flow.scripts.cylc_scan:main
-    cylc-scan-nt = cylc.flow.scripts.cylc_scan_nt:main
     cylc-search = cylc.flow.scripts.cylc_search:main
     cylc-set-verbosity = cylc.flow.scripts.cylc_set_verbosity:main
     cylc-show = cylc.flow.scripts.cylc_show:main
